AssetPreview Working!
Utilized Serialized Properties to get the element at a certain index of the ListView, and then FindPropertyRelative to get the correct variable in the element (the game object) to get the asset preview of it.
Also setup precaching of the images so that they load quickly when you first select the item.
Otherwise you have to reselect the object a couple of time for it to show up...